+++ CHANGELOG.md
@@ -45,12 +45,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 =======
 * Gmsh: read_file (.msh) and related methods, define_triangular_element_mesh, define_tetrahedron_element_mesh
+<<<<<<< HEAD
 * Circle2D: primitives (defined with 2 Arc2D)
+=======
+* Arc: reverse
+>>>>>>> fix_pt_belongs_arc2d
 *
 
 ### Fixed
 
 * Contour3D: average_center_point (use edge_polygon.points instead of points)
+* Arc2D: point_belongs (use is_trigo, consider point=start or end)
 *
 >>>>>>> origin/circle_primitives